# docs/api/arx_backend_api_spec.yaml
openapi: 3.0.3
info:
  title: Arx Backend API
  description: |
    Comprehensive API for the Arx Backend service providing BIM object management,
    asset inventory, CMMS integration, maintenance workflows, and compliance reporting.

    ## Authentication
    The API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Public endpoints: 100 requests/second, burst of 200
    - Authenticated endpoints: 50 requests/second, burst of 100

    ## Error Handling
    All endpoints return standard HTTP status codes with detailed error messages in JSON format.
  version: 1.0.0
  contact:
    name: Arx Development Team
    email: dev@arxos.com
  license:
    name: Proprietary
    url: https://arxos.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.arxos.com/api
    description: Production server
...
paths:
...
  /rooms:
    get:
      tags:
        - BIM Objects
      summary: List rooms
      description: Retrieve a paginated list of rooms
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: building_id
          in: query
          schema:
            type: integer
        - name: floor_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - BIM Objects
      summary: Create room
      description: Create a new room (admin/editor only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - building_id
                - floor_id
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                building_id:
                  type: integer
                floor_id:
                  type: integer
                room_number:
                  type: string
                room_type:
                  type: string
                area:
                  type: number
                  minimum: 0
                coordinates:
                  type: object
                  properties:
                    x:
                      type: number
                    y:
                      type: number
                    width:
                      type: number
                    height:
                      type: number
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
  /rooms/{id}:
    get:
      tags:
        - BIM Objects
      summary: Get room
      description: Retrieve a specific room by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'

    patch:
      tags:
        - BIM Objects
      summary: Update room
      description: Update an existing room (admin/editor only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                room_number:
                  type: string
                room_type:
                  type: string
                area:
                  type: number
                coordinates:
                  type: object
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'

    delete:
      tags:
        - BIM Objects
      summary: Delete room
      description: Delete a room (admin/editor only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Room deleted successfully

  # BIM Objects - Devices
...
  /routes:
    post:
      tags:
        - Routes
      summary: Create route
      description: Create a new route
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - start_point
                - end_point
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                start_point:
                  type: object
                  properties:
                    x:
                      type: number
                    y:
                      type: number
                    floor_id:
                      type: integer
                end_point:
                  type: object
                  properties:
                    x:
                      type: number
                    y:
                      type: number
                    floor_id:
                      type: integer
                route_type:
                  type: string
                  enum: [evacuation, maintenance, visitor, emergency]
                waypoints:
                  type: array
                  items:
                    type: object
                    properties:
                      x:
                        type: number
                      y:
                        type: number
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'

  # Object Info and SVG
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        coordinates:
          type: object
          additionalProperties: true
        properties:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [active, inactive, maintenance, archived]
        assigned_to:
          type: string
        locked:
          type: boolean
        locked_by:
          type: integer
        locked_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        start_point:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            floor_id:
              type: integer
        end_point:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            floor_id:
              type: integer
        route_type:
          type: string
          enum: [evacuation, maintenance, visitor, emergency]
        waypoints:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean